LayoutTests/imported/w3c/web-platform-tests/css/css-ui/parsing/webkit-appearance-alias-expected.txt


PASS -webkit-appearance is a legacy name alias for appearance
